Tracking student performance
This program keeps track of student performance in an arbitrary course. The first command-line
the argument to the program is a non-optional non-empty string representing the course name. The second argument, which
is optional, is a number between 0 and 10 inclusive that represents the number of assignments for the course. If the optional
argument is not provided, the program defaults to 5 assignments. We use n to represent the number of assignments (whether
specified as an argument or the default value of 5).